+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - `obsType`
   - `featureType`
   - `featureValueType`
+- Added the optional `coordinationValues` property for view definitions in the view config, as a shortcut to set coordination values for independent views. Used during the `initStrategy: 'auto'` initialization process.
 
 ### Changed
 - Fixed buggy view closing behavior by using the view `uid` rather than the index as the component `key`.